ask before delete

+++ inventory/src/Category.js
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from 'react'
 import axios from 'axios';
 import { Link } from 'react-router-dom';
+import { Dropdown, ButtonGroup } from 'react-bootstrap';
 
 function Category() {
   const [category, setCategory] = useState([])
@@ -11,15 +12,20 @@ function Category() {
 
   const handleDelete = async(id) => {
     try{
+      if (confirmDelete()){
       await axios.delete('http://localhost:8081/category/'+id)
       window.location.reload()
+      }
     }catch(err) {
       console.log(err)
     }
 
   }
 
- 
+  const confirmDelete = () => {
+    const isConfirmed = window.confirm('Are you sure you want to delete?');
+    return isConfirmed
+  }
 
   return (
     <div className='d-flex vh-100  justify-content-center align-items-center'>
@@ -34,7 +40,6 @@ function Category() {
                   <th>Name</th>
                   <th>Description</th>
                   <th>Action</th>
-                  <th>Action</th>
                 </tr>
               </thead>
               <tbody>
@@ -43,8 +48,18 @@ function Category() {
                     <tr key={i}>
                       <td>{data.name}</td>
                       <td>{data.description}</td>
-                      <td><Link to={`updateCategory/${data.id}`} className='btn btn-primary'>Update</Link></td>
-                      <td><button className='btn btn-danger ms-2' onClick={ e => handleDelete(data.id)}>Delete</button></td>
+                      <td>
+                      <ButtonGroup>
+                      <Link to={`updateCategory/${data.id}`} className='btn btn-light'>Update</Link>
+                      <Dropdown >
+                        <Dropdown.Toggle split variant="light" id="dropdown-split-basic" />
+                        <Dropdown.Menu>
+                          <Dropdown.Item onClick={e => handleDelete(data.id)}>Delete</Dropdown.Item>
+                         
+                        </Dropdown.Menu>
+                      </Dropdown>
+                    </ButtonGroup>
+                      </td>
                     </tr>
                   ))
                   }
