+++ inventory/src/Location.js
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from 'react'
 import axios from 'axios';
 import { Link } from 'react-router-dom';
+import { Dropdown, ButtonGroup } from 'react-bootstrap';
 
 function Location() {
   const [location, setLocation] = useState([])
@@ -11,12 +12,18 @@ function Location() {
 
   const handleDelete = async(id) => {
     try{
+      if (confirmDelete()){
       await axios.delete('http://localhost:8081/location/'+id)
       window.location.reload()
+      }
     }catch(err) {
       console.log(err)
     }
 
+  }
+  const confirmDelete = () => {
+    const isConfirmed = window.confirm('Are you sure you want to delete?');
+    return isConfirmed
   }
 
  
@@ -30,7 +37,6 @@ function Location() {
                 <tr>
                   <th>Name</th>
                   <th>Action</th>
-                  <th>Action</th>
                 </tr>
               </thead>
               <tbody>
@@ -38,8 +44,16 @@ function Location() {
                   location.map((data, i) => (
                     <tr key={i}>
                       <td>{data.name}</td>
-                      <td><Link to={`updateLocation/${data.id}`} className='btn btn-primary'>Update</Link></td>
-                      <td><button className='btn btn-danger ms-2' onClick={ e => handleDelete(data.id)}>Delete</button></td>
+                      <ButtonGroup>
+                      <Link to={`updateLocation/${data.id}`} className='btn btn-light'>Update</Link>
+                      <Dropdown >
+                        <Dropdown.Toggle split variant="light" id="dropdown-split-basic" />
+                        <Dropdown.Menu>
+                          <Dropdown.Item onClick={e => handleDelete(data.id)}>Delete</Dropdown.Item>
+                         
+                        </Dropdown.Menu>
+                      </Dropdown>
+                    </ButtonGroup>
                     </tr>
                   ))
                   }
