+++ inventory/src/Supplier.js
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from 'react'
 import axios from 'axios';
 import { Link } from 'react-router-dom';
+import { Dropdown, ButtonGroup } from 'react-bootstrap';
 
 function Supplier() {
   const [supplier, setSupplier] = useState([])
@@ -11,14 +12,19 @@ function Supplier() {
 
   const handleDelete = async(id) => {
     try{
+      if (confirmDelete()){
       await axios.delete('http://localhost:8081/supplier/'+id)
       window.location.reload()
+      }
     }catch(err) {
       console.log(err)
     }
 
   }
-
+  const confirmDelete = () => {
+    const isConfirmed = window.confirm('Are you sure you want to delete?');
+    return isConfirmed
+  }
  
 
   return (
@@ -32,7 +38,6 @@ function Supplier() {
                   <th>Address</th>
                   <th>Phone Number</th>
                   <th>Action</th>
-                  <th>Action</th>
                 </tr>
               </thead>
               <tbody>
@@ -42,8 +47,18 @@ function Supplier() {
                       <td>{data.name}</td>
                       <td>{data.address}</td>
                       <td>{data.phone}</td>
-                      <td><Link to={`updateSupplier/${data.id}`} className='btn btn-primary'>Update</Link></td>
-                      <td><button className='btn btn-danger ms-2' onClick={ e => handleDelete(data.id)}>Delete</button></td>
+                      <td>
+                      <ButtonGroup>
+                      <Link to={`updateSupplier/${data.id}`} className='btn btn-light'>Update</Link>
+                      <Dropdown >
+                        <Dropdown.Toggle split variant="light" id="dropdown-split-basic" />
+                        <Dropdown.Menu>
+                          <Dropdown.Item onClick={e => handleDelete(data.id)}>Delete</Dropdown.Item>
+                         
+                        </Dropdown.Menu>
+                      </Dropdown>
+                    </ButtonGroup>
+                      </td>
                     </tr>
                   ))
                   }
